moved traits to own file, updated Changelog
