+++ core_types/src/dna/mod.rs
@@ -28,6 +28,7 @@ pub mod capabilities;
 pub mod dna;
 pub mod entry_types;
 pub mod fn_declarations;
+pub mod traits;
 pub mod wasm;
 pub mod zome;
